requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
qdrant-client
numpy
rank-bm25
pytest
ruff
requests
langchain
langchain-openai
langchain-community
openai
beautifulsoup4
python-dotenv
lxml
PyPDF2
